Introduce a PAL
